Dragon folder restructuring
